Slash Command対応
Closes #48
