+++ go.mod
@@ -10,12 +10,11 @@ require (
 	github.com/ikawaha/kagome-dict/uni v1.1.9
 	github.com/ikawaha/kagome/v2 v2.9.5
 	github.com/nasa9084/go-switchbot/v3 v3.0.3
-	github.com/slack-go/slack v0.12.4
+	github.com/slack-go/slack v0.12.5
 
 )
 
 require (
-	github.com/alecthomas/participle/v2 v2.1.1 // indirect
 	github.com/google/uuid v1.6.0 // indirect
 	github.com/gorilla/websocket v1.5.1 // indirect
 	golang.org/x/net v0.21.0 // indirect
